refactor(diagnostics): split the lifecycle seams out of four oversized files
PROD-SIZE-1 failed on the Stage-0 commit: startup.rs (696), create.rs (755),
producer/record.rs (646) and producer/tests_support.rs (606) all crossed the
600-line cap, and all four were under it on main, so a ratchet entry would be a
net-new exception rather than recorded debt.

Each file splits along the seam the instrumentation created:

- runtime/startup_lifecycle.rs owns the observable agent start (the guard plus
  the terminal classification), leaving startup.rs at its pre-Stage-0 shape.
- service/create_lifecycle.rs owns the observable session create, the bounded
  origin label, and the terminal classification.
- producer/record/lifecycle.rs owns the lifecycle half of the record factory,
  so every rule about what a lifecycle record may carry sits in one module.
- producer/tests_collector_fixture.rs owns the hand-served loopback collector,
  which was already a distinct concern inside tests_support.rs.

No behavior change: the moved code is byte-identical apart from visibility and
imports.
